+++ js/core/event_listeners.ts
@@ -126,6 +126,7 @@ export function setupEventListeners(): void {
     setupIconAction('icon-audiostudio', () => launchKernelApp('audiostudio'));
     setupIconAction('icon-docexplorer', () => launchKernelApp('docexplorer'));
     setupIconAction('icon-mediaplayer', () => launchKernelApp('mediaplayer'));
+    setupIconAction('icon-voiceassistant', () => launchKernelApp('voiceassistant'));
     setupIconAction('icon-notepad', () => launchKernelApp('notepad'));
     setupIconAction('icon-paint', () => launchKernelApp('paint'));
     setupIconAction('icon-explorer', () => launchKernelApp('explorer'));